use Decimal() for quantity in send_order()

+++ CryptoBotV1.py
@@ -15,6 +15,7 @@ from binance.websockets import BinanceSocketManager
 from binance.enums import *
 # extra
 import math
+from decimal import Decimal
 import sys,os
 import json
 import datetime as dt
@@ -221,14 +222,14 @@ def process_m_message(msg):
                                 symbol.stop_loss = 0
                                 symbol.buy_price= current_price
                                 # take profit after average price
-                                order_succeeded = send_order(side=SIDE_BUY , quantity=symbol.piramidding_amount*symbol.margin_ratio, ticker=symbol.ticker,price=current_price,order_type=ORDER_TYPE_LIMIT,isolated=True,side_effect="MARGIN_BUY",timeInForce=TIME_IN_FORCE_GTC)
+                                order_succeeded = send_order(side=SIDE_BUY , quantity=Decimal(symbol.piramidding_amount*symbol.margin_ratio), ticker=symbol.ticker,price=current_price,order_type=ORDER_TYPE_LIMIT,isolated=True,side_effect="MARGIN_BUY",timeInForce=TIME_IN_FORCE_GTC)
                                 symbol.average_price = (symbol.amount*symbol.average_price + symbol.piramidding_amount*current_price) / (symbol.amount+symbol.piramidding_amount)
                                 symbol.take_profit = get_amount(symbol.average_price * 1.011,symbol.precision_minPrice, False)
                                 if order_succeeded:
                                     symbol.log_buy(amount=symbol.piramidding_amount ,buy_price=current_price, ticker=name, money=symbol.money)
                                     # get amount for limit sell order
                                     asset = client.get_isolated_margin_account(symbols=name)
-                                    symbol.amount = get_amount(float(asset["assets"][0]["baseAsset"]["free"]), symbol.precision)
+                                    symbol.amount = Decimal(get_amount(float(asset["assets"][0]["baseAsset"]["free"]), symbol.precision))
                                     take_profit_order = send_order(side=SIDE_SELL , quantity=symbol.amount, ticker=symbol.ticker,price=symbol.take_profit,order_type=ORDER_TYPE_LIMIT,isolated=True,side_effect="AUTO_REPAY",timeInForce=TIME_IN_FORCE_GTC)
                                     symbol.has_position = True
                                     order_succeeded = False
@@ -255,7 +256,7 @@ def process_m_message(msg):
                             print(f"symbol.amount: {type(symbol.amount)}")
                             print(f"symbol.margin_ratio: {type(symbol.margin_ratio)}")
                             symbol.buy_price= symbol.average_price
-                            order_succeeded = send_order(side=SIDE_BUY , quantity=symbol.amount*symbol.margin_ratio, ticker=symbol.ticker,price=symbol.average_price,order_type=ORDER_TYPE_LIMIT,isolated=True,side_effect="MARGIN_BUY",timeInForce=TIME_IN_FORCE_GTC)
+                            order_succeeded = send_order(side=SIDE_BUY , quantity=Decimal(symbol.amount*symbol.margin_ratio), ticker=symbol.ticker,price=symbol.average_price,order_type=ORDER_TYPE_LIMIT,isolated=True,side_effect="MARGIN_BUY",timeInForce=TIME_IN_FORCE_GTC)
                             print("order succeeded",order_succeeded)
 
                             if order_succeeded:
@@ -268,7 +269,7 @@ def process_m_message(msg):
 
                                 # get amount in wallet
                                 asset = client.get_isolated_margin_account(symbols=name)
-                                symbol.amount = get_amount(float(asset["assets"][0]["baseAsset"]["free"]), symbol.precision)
+                                symbol.amount = Decimal(get_amount(float(asset["assets"][0]["baseAsset"]["free"]), symbol.precision))
                                 take_profit_order = send_order(side=SIDE_SELL , quantity=symbol.amount, ticker=symbol.ticker,price=symbol.take_profit,order_type=ORDER_TYPE_LIMIT,isolated=True,side_effect="AUTO_REPAY",timeInForce=TIME_IN_FORCE_GTC)
                                 
                 
@@ -334,7 +335,7 @@ def callback_isolated_accounts(msg):
             symbol.has_position = True
 
             asset = client.get_isolated_margin_account(symbols=name)
-            symbol.amount = get_amount(float(asset["assets"][0]["baseAsset"]["free"]), symbol.precision)
+            symbol.amount = Decimal(get_amount(float(asset["assets"][0]["baseAsset"]["free"]), symbol.precision))
             take_profit_order = send_order(side=SIDE_SELL , quantity=symbol.amount, ticker=symbol.ticker,price=symbol.take_profit,order_type=ORDER_TYPE_LIMIT,isolated=True,side_effect="AUTO_REPAY",timeInForce=TIME_IN_FORCE_GTC)
